added option to remove custom networks (#2183)
* added option to remove custom networks
* Fix remaining long type issues

+++ app/src/main/java/com/alphawallet/app/entity/EthereumProtocolParser.java
@@ -89,7 +89,7 @@ public class EthereumProtocolParser
                                 readState = ParseState.READ_DIRECTIVE;
                                 break;
                             case CHAIN_ID:
-                                result.chainId = item.getValueBI().intValue();
+                                result.chainId = item.getValueBI().longValue();
                                 readState = ParseState.READ_DIRECTIVE;
                                 break;
                             case READ_TYPE:
